backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
cramjam==2.8.3
distlib==0.3.8
Django==4.1.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
filelock==3.13.3
googlemaps==4.10.0
idna==3.6
mongoengine==0.28.2
pipdeptree==2.16.2
platformdirs==4.2.0
pymongo==4.6.3
PyMySQL==1.1.0
python-dotenv==1.0.1
python-snappy==0.7.1
pytz==2024.1
requests==2.31.0
sqlparse==0.2.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
virtualenvwrapper-win==1.2.7
